Move to types/ directory, which allows imports of User
